[ALPS02076693]MMProfile: use inline dummy APIs
[Detail] the MMProfile need to be a switchable module.
1.If in the Makefile it defines "obj- += mmp/",
meanwhile in the Kconfig "CONFIG_MMPROFILE" is not set to "y",
the mmp driver is not linked into kernel. So we firstly put
mmprofile.h under drivers/misc/mediatek/mmp/, and in it use marco
to define dummy functions.
2.The dummy functions defined by marco will result in build errors
in some other kernel modules which include the header.

[Solution] change the marco definitions to "static inline" functions
and other kernel modules shall add include path
"drivers/misc/mediatek/mmp/", and include the header in source code
as "#include <mmprofile.h>".

[Feature] Others

MTK-Commit-Id: 8c223591fc779f49ab19408b05e9bbcf5d8aa9ae

Change-Id: I208c5d4a535fed53b50737326e49192d91f0a920
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02076693