[ALPS02092167] irqchip: add MediaTek intpol driver support
[Detail]
Mediatek SoCs have an GIC interrupt polarity extension which allows
to swap the polarity for given interrupts.

[Solution]
Support interrupt polarity setting using gic_arch_extn.
Because ARM64 doesnt have soc mach layer to install arch extension,
this is implemented as a dummy interrupt controller.

Due to depedency, this is not the implementation got accepted in
upstream.

[Feature] Others

MTK-Commit-Id: 5e391ad15af8dabedcbfc96ba389b64cd4f31ffe

Change-Id: I8e954c9dfe26bca27df4c83a39a4f537f4017291
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02092167