[ALPS02076693] MMProfile: fix kernel-3.18 build error
[Detail] there are build errors of MMProfile driver:
kernel-3.18/drivers/misc/mediatek/mmp/
The reason is, in drivers/misc/mediatek/Makefile, it was defined as
obj-() += mmp/
So when CONFIG_MTK_MMPROFILE_SUPPORT is not set =y, the mmp/ driver
(no matter its src/ or dummy/ implementation) was not linked into
kernel, and other module which used mmp/ APIs would result in
"undefined reference to..." build error.

[Solution] fix the build error in 2 aspects:
1. in drivers/misc/mediatek/Makefile, define as
obj-() += mmp/ (well use this configuration to
control MMProfile driver module.)
2. modfy the header files (mmprofile.h & mmprofile_internal.h)
so that when CONFIG_MMPROFILE is not set =y, dummy "static inline"
functions can be used as API functions replacement.

[Feature] Others

MTK-Commit-Id: 16872ea87d8319d3719bf0ce26fa168bcab53e7e

Change-Id: I72932bae0163dc5d65ce57343c037febcfcba0b9
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02076693