[ALPS02148714] genirq: backport hierarchy irq domain
[Detail]
genirq: Work around __irq_set_handler vs stacked domains ordering issues

Backport hierarchy irqdomains support from 3.19

[Solution]
With the introduction of stacked domains, we have the issue that,
depending on where in the stack this is called, __irq_set_handler
will succeed or fail: If this is called from the inner irqchip,
__irq_set_handler() will fail, as it will look at the outer domain
as the (desc->irq_data.chip == &no_irq_chip) test fails (we havent
set the top level yet).

This patch implements the following: "If there is at least one
valid irqchip in the domain, it will probably sort itself out".
This is clearly not ideal, but it is far less confusing then
crashing because the top-level domain is not up yet.

[ tglx: Added comment and a protection against chained interrupts in
that context ]

(cherry picked from commit f86eff222fab ("genirq: Work around
__irq_set_handler vs stacked domains ordering issues"))

[Feature] hierarchy irq domain

MTK-Commit-Id: 8036cf1dad03675d81f70f7a020d0b9d3c2fca9a

CR-Id: ALPS02148714
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Cc: Yingjoe Chen <yingjoe.chen@mediatek.com>
Cc: Bjorn Helgaas <bhelgaas@google.com>
Cc: linux-arm-kernel@lists.infradead.org
Cc: Jiang Liu <jiang.liu@linux.intel.com>
Link: http://lkml.kernel.org/r/1416048553-29289-3-git-send-email-marc.zyngier@arm.com
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
Change-Id: Idcb1368d7f880da1daaffcbede6450371996086d