[ALPS02323364] wmt/stp: Code optimization
[Detail]

1. Explicit null dereferenced
A null pointer dereference will occur.
In wmt_gpio_deinit: Dereference of an explicit null value (CWE-476)

2. Dead default in switch
The statements in the default case are never executed.
In _mt_combo_plt_do_deep_idle: Default in switch statement is dead
code (CWE-561)

[Solution]

1. Explicit null dereferenced
Assignment to move to the outside of the loop, avoid to cause
secondary access.

2. Dead default in switch
Remove not execute statement

[Feature] null

MTK-Commit-Id: 0abd3b45bbf540ca7b79e4866882854cc04b6702

Change-Id: Id37147fb954006fcb04fc3644a5c503ebf36718b
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02323364