[ALPS02316012] mem: adjust page fallback allocations
[Detail]
Memory fragmentation is a big issue during the long-term
quality tests. To ease this situation, we adjust the
behavior in the page fallback allocation.
Besides, we add a judgement in LMK driver to let it be
more aggressive when it hits the condition of vm_swap_full.

[Solution]
In mm, we merged the following patches from the kernel tree,
1. 3a1086fba92b6e2311b6a342f68bc380beb240fe
2. 9c0415eb8cbf0c8fd043b6c0f0354308ab099df5

In LMK driver, we added the following judgement,
"if (vm_swap_full()) /* Halve other_free */"

[Feature] MTBF(*)

MTK-Commit-Id: a0123ca17da3b4fb73853c2066e11332e50400e1

Change-Id: Ibf4c36cdcf88c329ca262f4c8ad0ea1a038f812d
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02316012