[ALPS02340309] TEEI: Fix factory mode issue
[Detail]
In factory mode, the SODI will enter ealier before TEEI is ready causing
HW reboot. Therefore, we add API for power module to judge if the device
can enter SODI/SODI3 or not. Only when TEEI is ready can the device enter
SODI/SODI3.

[Solution]
When SODI/SODI3 is about to enter, it will then use this API to check if
TEEI is ready. If not, it will not enter SODI.

[Feature] Factory Mode

MTK-Commit-Id: f3bead2d023f73ec3974e13b6816bd42bc317ab3

Change-Id: I38aa628df82a007192987f6739ff0b77afbda0f5
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02340309