[ALPS02344453] GPUDVFS: KE when usb unplugged
[Detail]
when usb unplugged with device, it would shut down univpll for
power saving, which we use this pll for temperary state during
the time for switching frequency. it would cause system abnormal.

[Solution]
replace UNIVPLL_D3 with SYSPLL_D3 instead, since SYSPLL do not
turn off til last step of device suspend.

we only need to modify dtsi for this setting.

[Feature] GPU DVFS

MTK-Commit-Id: 53689ed5677711cdcbf1c6643f8b5bfffd855d3f

Change-Id: Ia7056d9266f645ee4f1d460d12e49382c72d4bca
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02344453