[ALPS02344208] SMI: Add the fb_notify ops for smi variant
[Detail]
The pm_runtime suspond and resume could not be called when system
suspend resume, we need to add the fb notify suspend resume in
order to backup and restore the larb register. Or the m4u would not
be configured after system resume.

[Solution]
Add the fb notify suspend resume for smi.

[Feature]
SMI.

MTK-Commit-Id: f24122ce2e4b24758c02c9cfdd4b8342a95ef205

Change-Id: I5a04425141c22a6bcbc9e34f1412b3fadd4bdadb
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02344208