[ALPS02333873] mmc: core: Enable / disable re-tuning
[Detail] N/A

[Solution] N/A

[Feature] N/A

Enable re-tuning when tuning is executed and
disable re-tuning when card is no longer initialized.

In the case of SDIO suspend, the card can keep power.
In that case, re-tuning need not be disabled, but, if
a re-tuning timer is being used, ensure it is disabled
and assume that re-tuning will be needed upon resume
since it is not known how long the suspend will last.

Signed-off-by: Adrian Hunter <adrian.hunter@intel.com>
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
(cherry picked from commit 79d5a65aeea43920bf3ff60791f317570dd6f54f)

MTK-Commit-Id: 19cefd495ba02b901535e7d9b763d263acdde4f7

Change-Id: I4b638267444227694be0a708618741713a0f963b
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02333873