[ALPS02373260] wlan: use EtherType to find if frame is EtherType V2
[Detail]
1. Wlan TXD will use ENUM_PKT_802_3 to tag if the out going frame

is EtherType V2, driver should use value in *ETHER_HEADER_LEN to

judge if the frame is Ethernet V2.

2. more details:

https://en.wikipedia.org/wiki/EtherType

EtherType numbering generally starts from 0x0800.
In modern implementations of Ethernet, the field within the Ethernet
frame used to describe the EtherType also can be used to represent the
size of the payload of the Ethernet Frame. Historically,
depending on the type of Ethernet framing that was in use on an Ethernet
segment, both interpretations were simultaneously valid, leading to ambiguity.
Ethernet v2 framing considered these octets to represent EtherType
while the original IEEE 802.3 framing considered these octets to
represent the size of the payload in bytes.

In order to allow packets using Ethernet v2 framing and packets using the
IEEE 802.3 framing to be used on the same Ethernet segment,
a unifying standard (IEEE 802.3x-1997) was introduced that required that
EtherType values be greater than or equal to 1536 (0x0600). That value was
chosen because the maximum length (MTU) of the data field of an Ethernet 802.3
frame was 1500 bytes (0x05DC). Thus, values of 1500 (0x05DC) and below for
this field indicate that the field is used as the size of the payload of
the Ethernet Frame while values of 1536 and above indicate that the field
is used to represent EtherType.

[Solution]

Use value in *ETHER_HEADER_LEN to judge if the frame is Ethernet V2.

[Feature]

MTK-Commit-Id: 536c7db29afb0a7b28c0303800b1ce9e65422dc8

Change-Id: I8dcdf8d76020cf98e3b6d8c8b2bc346c7d562ba2
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02373260