[ALPS02222102] Thermal: Replace the SW filter with a HW filter
[Detail]
In Everest, the temperature rise are huge and various on different phones.
We would not be able to descide a temperature rise is reasonable or just a noise,
so we decided to use a HW filter which gets 4 samples, drops min and max ones,
then average rest of them.

Because getting more samples, we also need to adjust our thermal controller
to meet the 2ms polling requirement of Big core.

[Solution] Replace the SW filter with a HW filter

[Feature] Thermal Management

MTK-Commit-Id: 671360573d23ad74cde2fcbfd29bf264dcc471d2

Change-Id: Iad1afd781bdab5db50f32fd36ea0282561272fa7
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02222102