[ALPS02396167] mm/cma: fix memory reclaim behavior on zone-movable-cma
[Detail]
zone-movable-cma allows ZONE_MOVABLE to be full of CMA pages.
This will lead to fail zone_watermark_ok checks during background
memory reclaims. Besides, zone-movable-cma will be used for
memory-lowpower mechanism. It is not reasonable to reclaim it
if no or less reclaimable pages.

[Solution]
This patch can help system reduce the CPU loading of kswapd.
Without this patch, you may see the CPU loading of kswapd to be
around 50% or above.
It contains the following modifications,
1. Dont count free_cma in kswapd for ZONE_MOVABLE.
2. View ZONE_MOVABLE as balanced if no reclaimable pages
or "few free pages and few reclaimable ones" in it.

[Feature] [Android Default] Streaming

MTK-Commit-Id: 7c9a479c9d8b7cd976f52cd1d5544c0e869cb8af

Change-Id: Ic6b86b0a79d0b19075e73f3646b4da9955978654
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02396167