[ALPS02401893] scp: set dram allocation range
[Detail]
SCP can access dram via remap mechanism. but the remap range in SCP is fixed.
To gurantee the allocated dram that is in the access range of SCP,
it limits the allocated range in device tree.

[Solution]
To gurantee the allocated dram that is in the access range of SCP,
it limits the allocat range from 0x4000_0000~0x8FFF_FFFF in device tree.

[Feature] Tinysys

MTK-Commit-Id: 0277f128679595210cbc73005dc515640e9f3334

Change-Id: If36e88f2975dee311522a02578af973cb21d5a50
CR-Id: ALPS02401893
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>