[ALPS02407314] USB: scan media more times
[Detail]
Currently, SCSI driver only detects that the media
is presented at the first time. So if the the media
is mounted later, the SCSI driver does not know
there is a new media.
[Solution]
Does not return directly when no media is presented.
Retry 5 times and set the timeout as 30sec. So SCSI
driver can detect all media mounted during these retries.

[Feature] USB TYPE-C

MTK-Commit-Id: 3a554ca4118fd128a0acae33afcd504ce6e66dc6

Change-Id: I48015dc4f0614c6327fc69830ba9e0eaa2bfe865
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02407314