[ALPS02296522] TEEI: update patches
[Detail]
In previous patch, the TEEI is not very stable, and tester will
run into problem if we repeat the test for a long time. Now, this
patch change to use worker thread for SMC call instead of IPI.

[Solution]
Change the TEEI driver to use worker thread for SMC call.

MTK-Commit-Id: b634e94dc359bff43533ce04fb44fd6637ca7f4a

Change-Id: I0248da0b3e628903cd72f532abea388097d6775b
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02296522
Feature: Microtrust TEEI