[ALPS02415721] AUXADC: Channel13 make the system crash.
[Detail]
APMIXED_BASE is not used in mt8163.
So there is not APMIXED_BASE node in mt8163.dtsi.
When selecting channel13 in IMM_auxadc_GetOneChannelValue function,
It will call the APMIXED_BASE to use which causes the system crash.

[Solution]
Add a config to control that mt8163 doest use the APMIXED_BASE.

MTK-Commit-Id: a820a49bf07ec3bccf82dbe6646c3dd31c2e12c5

Change-Id: Ied3db91c0efc19cf4c5b03c80943c4c17c665e34
CR-Id:ALPS02415721
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
Feature: AUXADC