[ALPS02413602] arm/arm64: add a debug hook function for ioremap
Add a debug hook entry pointer in ioremap for debugging.
For mt6797 platform, the debug hook function is used to trap
ioremap to 0x1001AXXX register more the once. Hence to
fix the race condition issue of accessing 0x1001AXXXX.

Add a debug hook entry pointer in ioremap for debugging and
implement the debug function for mt6797.

MTK-Commit-Id: e92faa8088eed9e56e9bed385b61286766627065

Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
Feature: arm/arm64
Change-Id: I3efb54e2201e8f5bd1fa4e21cceb34178a853a74
CR-Id: ALPS02413602