[ALPS02283088] Thermal: fix a change thermal policy bug
[Detail]
When default thermal config is in use and ATM is activated,
ATM coolers are not deactivated on changing thermal policy to ht120.
The reason is that exit point of ATM coolers is not reached.
Since changing thermal config is intended, ATM coolers should be deactivated
even when their exit points not reached.

[Solution]
Invoke coolers set_cur_state ops function to deactivate it instead of
calling mtk_cooling_wrapper_set_cur_state in which exit point is checked.

MTK-Commit-Id: 7e7d72ce0400434502e99236cddd3858562cf7bd

Change-Id: Ia26153bb50542a98cdb997cb22be1212b6b37414
CR-Id: ALPS02283088
Feature: Thermal Management
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>