[ALPS02377737] USB and UART Switch: Enable the feature for Everest.
[Detail]
The switch control on MT6797 has been changed from
UART to GPIO. The uart log will not be shown if the
value is not set correctly in GPIO MISC control register.

In preloader, the USB register base address is not correct.
Therefore it cannot get Line State and detect the cable plugged-in.

[Solution]
1. Change the switch control from UART to GPIO.
2. Fix USB register base address.
3. Enable the feature.

MTK-Commit-Id: 6be8a15fee04223440094f1dcd6b940f543bd586

Change-Id: I27f021a1414a64389af5a109eb4c588174edd4ca
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02377737
Feature: USB And UART Switch