[ALPS02444967] usb: fix xhci cant work after suspend
[Detail]
When system is suspended, if we plug in A-cable and as host.
the xhci host controller cant work well.
Because we use dynamic load xhci driver.
If we plugin A-cable, we will load xhci and the system resume will
call xhci xhci_plat_resume. It will restore our xhci setting.

[Solution]
We dont need xhci_plat_pm_ops in out otg design, and remove
it on CONFIG_USB_XHCI_MTK that is defined

MTK-Commit-Id: 2ff6c2e0d28967cde10e9fbccd6dfe0a79f79c92

Change-Id: I98356b2cb338b8672c3c7df7b4803bd76ce6a3a9
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02444967
Feature: [Android Default] USB OTG