[ALPS02408966] cpufreq: create symlink for cpu0 when cpu0 down
[Detail]
Most APPs assume cpu0 is always online. But for AMP architecture, it is
not true. If cpu0 down and the APP accesses the sysfs of cpu0, the
function may be abnormal.

[Solution]
If cpu0 down, create a symlink to other online cpu for cpu0.

MTK-Commit-Id: 9678a8703d0ffd70640fbe1fbf42118c449d23be

Change-Id: I5d166175565756b354818f210809595d1ef08d1d
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02408966
Feature: DVFS (Dynamic Voltage and Frequency Scaling)