[ALPS02438541] Hotspot: Everest 5G hotspot not support HT40
[Detail]
Gen3 driver uses prAdapter->rWifiVar.ucApBandwidth as the max BW limit of hotspot,
which is 20M by default, and can be configured in wifi.cfg.

[Solution]
To distinguish the hotspot BW setting on 2.4G and 5G, use two variables ucAp2gBandwidth and ucAp5gBandwidth
to represent respectively.
The default value should be configured to 20M on 2.4G band and 40M on 5G band.

MTK-Commit-Id: 3d3e9cde72dcdda09d0cd6fb063205c8e1b6e648

Change-Id: I2d7391fe306bf765c0c7883bf20da2395d33252c
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02438541
Feature: [Android Default] Wi-Fi Driver for SoC