[ALPS02430263] devapc driver: remove the usage of sscanf
[Detail]
The original implementation of "devapc_dbg_write" should not have
security vulnerability because it already has the boundary check.
However, to enhance the security, we add one more check and remove
the usage of sscanf to avoid false positive by code scanner.

[Solution]
Remove the use of sscanf to reduce the risk of security issue.

MTK-Commit-Id: 320be30ccb9f84808fefb75d5a55fa24fcdfe57a

Change-Id: I3d3aee60bcd047c8771e596fcd376282acf6933b
CR-Id: ALPS02430263
Feature: Others
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>