[ALPS02406790] mtd: m25p80: allow arbitrary OF match "jedec,spi-nor"
mtd: m25p80: allow arbitrary OF matching for "jedec,spi-nor"

When we added the "jedec,spi-nor" compatible string for use in this
driver, we added it as a modalias option. The modalias can be derived in
different ways for platform devices vs. device tree (of_*) matching. But
for device tree matching (the primary target of this identifier string),
the modalias is determined from the first entry in the compatible
property. IOW, the following properties would bind to this driver:

// Option (a), modalias = "spi-nor"
compatible = "jedec,spi-nor";

// Option (b), modalias = "spi-nor"
compatible = "idontknowwhatimdoing,spi-nor";

But the following would not:

// Option (c), modalias = "shinynewdevice"
compatible = "myvendor,shinynewdevice", "jedec,spi-nor";

So, wed like to match (a) and (c) (even when we dont have an explicit
entry for "shinynewdevice"), and wed rather not allow (b).

To do this, we
(1) always (for devices without specific platform data) pass the
modalias to the spi-nor library;
(2) rework the spi-nor library to not reject "bad" names, and
instead just fall back to autodetection; and
(3) add the .of_match_table to properly catch all "jedec,spi-nor".

This allows (a) and (c) without warnings, and rejects (b).

Signed-off-by: Brian Norris <computersforpeace@gmail.com>
(cherry picked from commit 43163022927b6e7d202a7e6f939c3f392465494d)

MTK-Commit-Id: 00fbe78084d9af8634333263b65be475aa4024b7

Change-Id: I421d9daf5958d39fe8a7d0785e565f9dae994785
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02406790
Feature: Others