[ALPS02367881] M4U : move pte alloc out of the lock
[Detail]
MTK add the heap_ops .add_freelist in ION to dealloc m4u mva
before shrink to avoid dead lock, currently we are going to
remove the .add_freelist ops in ION. So we need to move the
kmem_cache_alloc out the lock, or there maybe deadlock in the
following sequence:
a, ion_mm_heap_phys
-> acquire bufferInfo->lock
->m4u_alloc_mva
-> acquire m4u pgtable lock
-> kmem_cache_alloc(GFP_KERNEL) might sleep
b, shrik
-> ion_shrink
-> free ion deferred free list
->ion_buffer_destroy
->acquire bufferInfo->lock
->m4u_dealloc_mva
-> acuqire m4u pgtable lock
bufferInfo->lock in a is not the same with bufferInfo->lock in b.
If kmem_cache_alloc sleeps, then there is a deadlock.

[Solution] move pte alloc out of the lock

MTK-Commit-Id: 13f968ac107df68251e1039e83d808dcc3fdd8b1

Feature: Others
Change-Id: Icc4f66dc10a121e7226161a6a1ea35b313139c70
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02367881