[ALPS02485579] DISP: sync DSI1s operations with DSI0s
[Detail]
- In DSI Dual port case, it should synchronize DSI1s operations with
DSI0s to keep the both engines in the same state. If not, it may
cause the engine hang due to issuing command to engine while engine
is still busy.

[Solution]
- Set the commands of the both engines first. After that is done,
it can start the engines.
- Add polling DSI busy state in function DSI_WaitForNotBusy().
Originally, it waits event to check if DSI is busy or not. But it
may have chance that it is interrupted by other signal, and this
leads to not waiting DSIs state changing from busy to idle. In
the meanwhile, issuing a command to DSI engine while it is still
in the busy state will cause DSI engine to get stuck.

MTK-Commit-Id: 24062077861bcc9498d4b8a2b5fb48507dba0fe9

Change-Id: I7a4c0b15ffcf2138901e165b2c08483c6e8c1c2a
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02485579
Feature: Display Driver (DSI Interface)