[ALPS01562944] IPOH: check if current task is exiting
[Detail]
This is to fix a bug while trying to access /dev/snapshot.
The scenario is to create task to access this device node
and die suddenly. In releasing /dev/snapshot, the pm notifier
will be called to do all the things to be done in
POST_HIBERNATION stage, including set_env(). But it will try
to write block device in hibernation flow while current->fs
is already destroyed.

[Solution]
Check the task state in the notifier. They should only be called
in IPOH flow

MTK-Commit-Id: f5af684b09dfe485045c7887a9a9e3c1cf4d5160

Change-Id: I0bf4c6484dc5b55a4e5c6310f054a5a24c401eb8
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS01562944
Feature: Instant Power On