[ALPS02506690] USB:f_fs, Fail stale read IOs after disconnect
[Detail]
After a USB disconnect, endpoints for adb are disabled.
After this no IO is allowed on the endpoints.
Since, adbd is not aware of this disconnect, it may
still perform read/writes IO. For adb writes, IOs are
failed, but for adb reads kernel waits until endpoints
are enabled.

When a USB disconnect and adb read still queued
a buffer to kernel, ffs_epfile_io simply waits for
endpoint to be enabled. A next connect happens
and endpoints are enabled after set_alt, the adb
read stale buffer from previous session continues
and queues to endpoint.

All this time, adb did not close the epfile because
it did not get return status on the IOs which it
queued. This is an issue, because a new session
is not established and both userspace and kernel
goes out of sync.

To fix this issue, when endpoints are disbled
set epfile error. This epfile error is only cleared
in epfile open. This will ensure that after a USB
disconnect and connect, new session is established.

Also, return ENODEV if endpoints not enabled rather
than EINTR as EINTR case, and simply retries the
request. Incase usb_ep_queue failed, return -EIO
inspite of depend on return status from usb_ep_queue.

[Solution]

patch from internet
https://www.codeaurora.org/cgit/quic/la/kernel/msm-3.10/commit/?h=LA.BF.2.1.2_rb1.2&id=2311fb05f1543c58a99d6ad112054fc35db84a2f

MTK-Commit-Id: 100686c66396a97e21acaa72be9da23cca8badc0

Change-Id: Ieb5404a4ac2e1ec1a55cdd2cb0b063bfea94b14f
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02506690
Feature: UMS (Usb Mass Storage)