[ALPS02259367] clocksource: arch_timer: turn off virtual timer via dts
[Detail]
Allow the device tree to specify uninitialized timer registers

[Solution]
Some 32-bit (ARMv7) systems are architected like this:

* The firmware doesnt know and doesnt care about hypervisor mode and
we dont want to add the complexity of hypervisor there.

* The firmware isnt involved in SMP bringup or resume.

* The ARCH timer come up with an uninitialized offset (CNTVOFF)
between the virtual and physical counters.  Each core gets a
different random offset.

* The device boots in "Secure SVC" mode.

* Nothing has touched the reset value of CNTHCTL.PL1PCEN or
CNTHCTL.PL1PCTEN (both default to 1 at reset)

On systems like the above, it doesnt make sense to use the virtual
counter.  Theres nobody managing the offset and each time a core goes
down and comes back up it will get reinitialized to some other random
value.

This adds an optional property which can inform the kernel of this
situation, and firmware is free to remove the property if it is going
to initialize the CNTVOFF registers when each CPU comes out of reset.

Currently, the best course of action in this case is to use the
physical timer, which is why it is important that CNTHCTL hasnt been
changed from its reset value and its a reasonable assumption given
that the firmware has never entered HYP mode.

Note that its been said that on ARMv8 systems the firmware and
kernel really cant be architected as described above.  That means
using the physical timer like this really only makes sense for ARMv7
systems.

MTK-Commit-Id: f53a162b91b55a90c2c1d0a6636abb9fcab3d3db

cherry picked from
commit 65b5732d241b ("arch_timer: Allow the device tree to specify ...")
Feature: arm arch timer
Change-Id: I77f5604ebcba0031ac08dd619c78eb06cdd9f9e4
CR-Id: ALPS02259367
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>