Allow building if the kernel isn't in the top level directory of AOSP sources
In the Linaro repositories, we check out multiple kernels (one for every
supported board that requires a non-mainline kernel), so we prefer
putting kernels into a subdirectory structure similar to the device/
directory structure.
This patch allows building in that structure without affecting builds
in the traditional location for x20 builds.

Signed-off-by: Bernhard Rosenkränzer <Bernhard.Rosenkranzer@linaro.org>