build: Set -Wno-error on a number of new gcc5 and gcc6 diagnostics
Currently this kernel has been extended out-of-tree to make liberal use
of -Werror, this means it cannot be built using gcc5 or gcc6, both of
which add new diagnostic messages.

Fix this by finding the -Werror that caused the build to fail and add
a -Wno-error= targetting the new gcc warning.

I did consider fixing the warnings themselves but some hierarchies
(especially drivers/misc/mediatek) contain so many warnings that fixing
them is likely to cause nasty merge conflicts whenever the vendor
kernel is updated. This is doubly true for the various broken
indentation warnings!

Signed-off-by: Daniel Thompson <daniel.thompson@linaro.org>